Move functions into view models
A lot of functions lived in views and models that were
really view model-type functions.
